moved things to init.py to reduce infinite loop error
